+++ CHANGELOG.md
@@ -33,6 +33,7 @@
     * Removed additional padding from modals that were incorrectly inheriting it
     * Added missing responsive bootstrap classes required for Joomla editing views (#684)
     * Logo Particle now properly redirects to Home even if clicked from a subpage (#676)
+    * Menu Particle: max levels does not work (#698)
 
 # 5.0.1
 ## 07/16/2015